+++ app/Http/Controllers/backend/taskCategoryController.php
@@ -11,12 +11,14 @@ class taskCategoryController extends Controller
 {
     public function index()
     {
+        $this->authorize('create_task_category');
         $taskCategory = auth()->user()->taskCategorys;
         return view('backend.page.task.category',compact('taskCategory'));
     }
 
     public function store(taskCategoryRequest $request)
     {
+        $this->authorize('create_task_category');
         auth()->user()->taskCategorys()->create([
             'name' => $request->name,
             'description' => $request->description,
@@ -26,7 +28,11 @@ class taskCategoryController extends Controller
 
     public function update(taskCategoryRequest $request, Taskcategory $category)
     {
-        //$update_category = Taskcategory::findOrFail($id);
+        $this->authorize('create_task_category');
+        if (!auth()->id() === $category->user_id)
+        {
+            abort('403');
+        }
         $category->update([
             'name' => $request->name,
             'description' => $request->description,
@@ -36,6 +42,11 @@ class taskCategoryController extends Controller
 
     public function destroy(Taskcategory $category)
     {
+        $this->authorize('create_task_category');
+        if (!auth()->id() === $category->user_id)
+        {
+            abort('403');
+        }
         foreach ($category->tasks as $tasks){
             $tasks->delete();
         }
